Typescript bootstrapped. Nothing quite works yet
