--- bot/commandDescriptions.txt
@@ -1,28 +1,0 @@
-valimo - Aalto Valimo menu
-vm5 - AaltoVM5 menu
-alvari - Alvari menu
-portaali - Cafe Portaali menu
-cafespa - Cafespa menu
-chydenia - Chydenia menu
-electra - Electra menu
-gaudeamus - Gaudeamus Kirja & Kahvi menu
-hanken - Hanken menu
-osakunta - Hämäläis-Osakunta menu
-konetekniikka - Konetekniikka menu
-kvarkki - Kvarkki menu
-olivia - Olivia menu
-rafla - Rafla menu
-fredrika - Ravintola Fredrika Päärakennus menu
-serpens - Ravintola Serpens menu
-soc - Soc&Kom menu
-sahkotekniikka - Sähkötekniikka menu
-tuas - TUAS menu
-tietotekniikantalo - Tietotekniikantalo menu
-tietotie - Tietotie 6 menu
-taffa - Täffä menu
-unicafe_metsatalo - UniCafe Metsätalo menu
-unicafe_paarakennus - UniCafe Päärakennus menu
-unicafe_topelias - UniCafe Topelias menu
-unicafe_ylioppilasaukio - UniCafe Ylioppilasaukio menu
-help - get the commands of the bot
-feedback - send feedback to the developers<3
