--- bot/help.txt
@@ -1,5 +1,0 @@
-Hello!
-
-Send me your location, type /help or simply enter / to the message box to view commands.
-
-Powered by kanttiinit.fi
