--- bot/start.txt
@@ -1,2 +1,0 @@
-Type / to get started
-or /help to see what I do.
