+++ index.js
@@ -1,13 +1,13 @@
 const express = require('express');
-const packageInfo = require('./package.json');
 const bodyParser = require('body-parser');
 const cors = require('cors');
+const packageInfo = require('./package.json');
 const bot = require('./bot');
 
 const app = express();
 
-process.on('unhandledRejection', err => {
-	bot.sendMessage(feedbackChat, '#SRS_ERROR: ' + err);
+process.on('unhandledRejection', (err) => {
+  bot.sendMessage(feedbackChat, `#SRS_ERROR: ${err}`);
 });
 
 const feedbackChat = process.env.CHAT_ID;
@@ -15,17 +15,17 @@ const feedbackChat = process.env.CHAT_ID;
 app.use(bodyParser.json());
 
 app
-.get('/', (req, res) => {
-	res.json({version: packageInfo.version});
-})
-.options('/feedback', cors())
-.post('/feedback', cors(), (req, res) => {
-	bot.sendMessage(feedbackChat, '#FEEDBACK_WEB:\n' + req.body.message);
-	res.json({success: true});
-})
-.post('/' + bot.token, (req, res) => {
-	bot.processUpdate(req.body);
-	res.sendStatus(200);
-});
+  .get('/', (req, res) => {
+    res.json({ version: packageInfo.version });
+  })
+  .options('/feedback', cors())
+  .post('/feedback', cors(), (req, res) => {
+    bot.sendMessage(feedbackChat, `#FEEDBACK_WEB:\n${req.body.message}`);
+    res.json({ success: true });
+  })
+  .post(`/${bot.token}`, (req, res) => {
+    bot.processUpdate(req.body);
+    res.sendStatus(200);
+  });
 
 app.listen(process.env.PORT || 3000);
