+++ package.json
@@ -4,21 +4,30 @@
   "description": "Kanttiinit bot: Lunch menus within telegram",
   "main": "index.js",
   "scripts": {
-    "start": "node index.js"
+    "build": "tsc",
+    "start": "node build/index.js"
   },
   "engines": {
-    "node": "6.0.0"
+    "node": "10.9.0"
   },
   "author": "Laged",
   "license": "BSD-2-Clause",
+  "devDependencies": {
+    "@types/node": "^10.12.18",
+    "eslint": "^5.12.0",
+    "eslint-config-airbnb-base": "^13.1.0",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-typescript": "^0.14.0",
+    "typescript": "^3.2.2",
+    "typescript-eslint-parser": "^21.0.2"
+  },
   "dependencies": {
-    "body-parser": "^1.15.0",
-    "cors": "^2.7.1",
-    "express": "~4.13.4",
-    "fs": "0.0.2",
-    "moment": "^2.13.0",
-    "node-fetch": "^1.5.1",
-    "node-telegram-bot-api": "~0.21.1",
-    "request": "^2.72.0"
+    "body-parser": "^1.18.3",
+    "cors": "^2.8.5",
+    "express": "^4.16.4",
+    "moment": "^2.23.0",
+    "node-fetch": "^2.3.0",
+    "node-telegram-bot-api": "^0.30.0",
+    "request": "^2.88.0"
   }
 }
