+++ src/commandDescriptions.json
@@ -1,0 +1,3 @@
+{
+  "message": "valimo - Aalto Valimo menu\n  vm5 - AaltoVM5 menu\n  alvari - Alvari menu\n  portaali - Cafe Portaali menu\n  cafespa - Cafespa menu\n  chydenia - Chydenia menu\n  electra - Electra menu\n  gaudeamus - Gaudeamus Kirja & Kahvi menu\n  hanken - Hanken menu\n  osakunta - Hämäläis-Osakunta menu\n  konetekniikka - Konetekniikka menu\n  kvarkki - Kvarkki menu\n  olivia - Olivia menu\n  rafla - Rafla menu\n  fredrika - Ravintola Fredrika Päärakennus menu\n  serpens - Ravintola Serpens menu\n  soc - Soc&Kom menu\n  sahkotekniikka - Sähkötekniikka menu\n  tuas - TUAS menu\n  tietotekniikantalo - Tietotekniikantalo menu\n  tietotie - Tietotie 6 menu\n  taffa - Täffä menu\n  unicafe_metsatalo - UniCafe Metsätalo menu\n  unicafe_paarakennus - UniCafe Päärakennus menu\n  unicafe_topelias - UniCafe Topelias menu\n  unicafe_ylioppilasaukio - UniCafe Ylioppilasaukio menu\n  help - get the commands of the bot\n  feedback - send feedback to the developers<3\n  "
+}
