+++ src/commands.json
@@ -1,0 +1,3 @@
+{
+  "message": "The following commands are available: <send location> the closest 3 open restaurants to you /<restaurantname> text of lunch at the given restaurant /food interactive help for deciding where to eat /sub posts the subway of the day /txt [restaurantName OR ID] text of lunch at the given restaurant /img [restaurantName OR ID] image of lunch at the given restaurant /restaurants a list of restaurantNames and IDs /start a greeting /help info about me /feedback [text] sends a message with FRIENDLY feedback to the developers ;) "
+}
