+++ src/help.json
@@ -1,0 +1,3 @@
+{
+  "message": "Hello! Send me your location, type /help or simply enter / to the message box to view commands. Powered by kanttiinit.fi"
+}
