+++ src/start.json
@@ -1,0 +1,3 @@
+{
+  "message": "Type / to get started or /help to see what I do."
+}
